GhostPDF - fix a couple more compiler warnings
The cast in pdf_fapi_get_float() just matches the one a couple of lines
below and I'm sure was missed accidetnally.

I'm not so sure about the one in ps_font_def_func() but it silences the
compiler.

+++ pdf/pdf_fapi.c
@@ -550,7 +550,7 @@ pdfi_fapi_get_float(gs_fapi_font *ff, gs_fapi_font_feature var_id, int index, fl
                     break;
                 }
                 if (v->type == PDF_INT) {
-                    *ret = v->value.i;
+                    *ret = (float)v->value.i;
                 }
                 else {
                     *ret = (float)v->value.d;
